refactor(entity-currency-fiat): rename registry to utils
