+++ domain/entity/currency-fiat/src/index.ts
@@ -1,6 +1,6 @@
 export * from "./schema";
 export * from "./define";
-export * from "./registry";
+export * from "./utils";
 export * from "./constants";
 export * from "./types";
 export * from "./slice";
