+++ domain/entity/currency-fiat/src/utils.ts
